pyproject.toml
[project]
name = "bot-hyperliquide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hyperliquid-python-sdk>=0.22.0",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.2.2",
]
